Mostly Readability and compiler warnings dealt with.

+++ trunk/ saving-face --username 1900Geek@gmail.com/src/sf_model_builder.cpp
@@ -15,7 +15,7 @@ namespace SF
 	{
 		time_t t;
 		time(&t);
-		return t;
+		return (long)t;
 	}
 
 	SF_MUID sf_model_builder::addNewModel()
